第 29 轮回归成果：权威 md 949→901（累计 3324→901）
- guild 315→274（本轮最大收益，-41）；dbmw 227→223、statics 70→67
- 剩余：guild 274、monitor 328、dbmw 223、statics 67、manager 5、coserver 3、community 1(REMAIN)
- 6 个服务已全清；纯 C++ 形态修复（禁止硬套 asm 规矩维持）

+++ dnf_decompile/source/DNFServer/GameServer/Monitor/DNFManagerServer.cpp
@@ -262,7 +262,7 @@ void CTcpManagerServer::SendTcpPacket(PacketHeader* pkt)
     char* buf = makePacketHeader(*(unsigned short*)pkt, ((RA_U16<2>*)pkt)->v);
     if (buf != 0)
     {
-        memcpy(buf + 10, (char*)pkt + 10, ((RA_U16<2>*)pkt)->v - 10);
+        memcpy(buf + 10, (char*)pkt + 10, (int)(((RA_U16<2>*)pkt)->v - 10));
         SendToServer(buf);
     }
 }
